requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

python-dotenv

# Linting and formatting
ruff==0.5.3


# Core Machine Learning Frameworks (CPU-only)
torch==2.7.1
# torchvision==0.22.1
# torchaudio==2.7.1


# Sentence Embeddings
sentence-transformers==5.0.0

# Core RAG and LangChain Framework
langchain==0.3.27
langchain-community==0.3.27

# Vector Database Client
# Pin to stable version to avoid v4.16.0-4.16.3 bugs
weaviate-client==4.16.6

# Data Processing and File Parsing
unstructured[pdf]==0.14.8
langdetect

# Image processing
pillow>=11.1.0

# Web and Network
httpx
streamlit==1.47.0

# Utilities
protobuf==5.29.0
grpcio==1.63.0
grpcio-health-checking==1.63.0
grpcio-status==1.63.0

# CLI and logging
rich>=13,<14

# for Windows (but doesn't hurt on Linux)
# https://github.com/Unstructured-IO/unstructured/issues/2237
# python-magic-bin==0.4.14
